+++ CHANGELOG.md
@@ -30,6 +30,24 @@
 
 ## Unreleased
 
+### Added
+- Kara status: global status to run, pause or stop the karaoke.
+  - play: Same as previous behaviour.
+  - pause: No additional song is played by the player. The player finishes playing current song.
+  - stop: Player stops playing, playlist is emptied, can't add song to playlist.
+- Kara status to digest route.
+
+### Changed
+- Playlist routes:
+  - `/player/*` > `/playlist/device/*`
+  - `/playlist/` > `/playlist/entries/`
+  - `/playlist/<id>/` > `/playlist/entries/<id>/`
+  - `/playlist/player/` > `/playlist/digest/`
+- Digest route (old player route) content change:
+  - `status` > `player_status`
+  - `manage` > `player_manage`
+  - `errors` > `player_errors`
+
 ## 1.1.0 - 2018-01-25
 
 ### Added